chore: created UserService
